Move session logic to hook for easier testing
